Move audit DbContext classes to Audit module
Relocates IAuditDbContext and AuditDbContext from the EntityFramework
module to the Audit module for better architectural alignment.

Changes:
- Move src/EntityFramework/IAuditDbContext.cs → src/Audit/IAuditDbContext.cs
- Move src/EntityFramework/AuditDbContext.cs → src/Audit/AuditDbContext.cs
- Update namespace from Wangkanai.EntityFramework to Wangkanai.Audit

These classes are audit-specific abstractions that belong with other
audit domain types rather than general EF utilities.
